requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
aioredis==1.3.1
    # via channels-redis
amqp==5.1.1
    # via kombu
anyio==3.6.2
    # via watchgod
asgiref==3.6.0
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   uvicorn
async-timeout==4.0.2
    # via
    #   aioredis
    #   redis
attrs==23.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==23.1.2
    # via daphne
automat==22.10.0
    # via twisted
beautifulsoup4==4.12.2
    # via -r requirements/./base.in
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements/./base.in
    #   django-celery-beat
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
channels==3.0.5
    # via channels-redis
channels-redis==3.4.0
    # via -r requirements/./base.in
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cron-descriptor==1.2.35
    # via django-celery-beat
cryptography==40.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
deprecated==1.2.13
    # via redis
django==4.2
    # via
    #   -r requirements/./base.in
    #   channels
    #   django-celery-beat
    #   django-cors-headers
    #   django-model-utils
    #   django-redis
    #   djangorestframework
    #   drf-yasg
django-celery-beat==2.5.0
    # via -r requirements/./base.in
django-cors-headers==3.11.0
    # via -r requirements/./base.in
django-model-utils==4.3.1
    # via -r requirements/./base.in
django-redis==5.2.0
    # via -r requirements/./base.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/./base.in
    #   drf-yasg
drf-yasg==1.21.5
    # via -r requirements/./base.in
environs==9.5.0
    # via -r requirements/./base.in
flake8==6.0.0
    # via -r requirements/dev.in
gunicorn==20.1.0
    # via -r requirements/./base.in
h11==0.14.0
    # via uvicorn
hiredis==2.2.2
    # via aioredis
httptools==0.5.0
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
kombu==5.2.4
    # via celery
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via environs
mccabe==0.7.0
    # via flake8
msgpack==1.0.5
    # via channels-redis
packaging==23.1
    # via
    #   drf-yasg
    #   marshmallow
    #   redis
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/./base.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pyopenssl==23.1.1
    # via twisted
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-dotenv==1.0.0
    # via
    #   environs
    #   uvicorn
pytz==2023.3
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via uvicorn
redis==4.3.1
    # via
    #   -r requirements/./base.in
    #   django-redis
requests==2.29.0
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   python-dateutil
    #   service-identity
sniffio==1.3.0
    # via anyio
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
twisted[tls]==22.10.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.5.0
    # via twisted
tzdata==2023.3
    # via django-celery-beat
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.15
    # via requests
uvicorn[standard]==0.17.6
    # via -r requirements/./base.in
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchgod==0.8.2
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.2
    # via uvicorn
whitenoise==6.0.0
    # via -r requirements/./base.in
wrapt==1.15.0
    # via deprecated
zope-interface==6.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
